Add a modal confirmation to confirm program deletion

+++ client/components/dashboard/dashboard.tsx
@@ -1,4 +1,4 @@
-import { ReactElement } from 'react';
+import { ReactElement, useState } from 'react';
 import { Navigate, useNavigate } from 'react-router-dom';
 import { useAuthContext } from '../../hooks/use-auth-context';
 import usePrograms from '../../hooks/use-programs';
@@ -6,20 +6,36 @@ import styles from './dashboard.module.css';
 import * as ProgramsService from '../../services/programs-service';
 import ProgramsTable from '../programs-table/programs-table';
 import Button from '../form-controls/button';
+import Modal from '../modal/modal';
+import Confirmation from '../../confirmation/confirmation';
+import { ProgramData } from '../../../common/api-types';
 
 export default function Dashboard(): ReactElement {
   const { user, userHasLoaded } = useAuthContext();
   const { programs, setPrograms } = usePrograms();
+  const [programToDelete, setProgramToDelete] = useState<ProgramData | null>(null);
   const navigate = useNavigate();
 
   if (userHasLoaded && !user) {
     return <Navigate to='/auth/log-in' replace={true} />;
   }
 
-  async function handleDelete(programId: string) {
+  async function handleDelete() {
+    if (!programToDelete) return;
+
+    const programId = programToDelete.id;
     const result = await ProgramsService.deleteProgram(programId);
+
     if (!result) return;
+
     setPrograms(programs.filter(p => p.id !== programId));
+    setProgramToDelete(null);
+  }
+
+  function handleDeleteButtonPressed(programId: string) {
+    const program = programs.find(p => p.id === programId);
+    if (!program) return;
+    setProgramToDelete(program);
   }
 
   function handleEdit(programId: string) {
@@ -33,7 +49,11 @@ export default function Dashboard(): ReactElement {
           <h2 className={styles.heading}>Your Programs</h2>
         </div>
         {programs.length !== 0
-          ? <ProgramsTable programs={programs} handleDelete={handleDelete} handleEdit={handleEdit} />
+          ? <ProgramsTable
+            programs={programs}
+            handleDelete={handleDeleteButtonPressed}
+            handleEdit={handleEdit}
+          />
           : <p>You don't have any programs yet.</p>}
         <Button
           className={styles.newProgramButton}
@@ -42,6 +62,14 @@ export default function Dashboard(): ReactElement {
           New Program
         </Button>
       </section>
+      <Modal open={!!programToDelete}>
+        <Confirmation
+          message={`Are you sure you want to delete "${programToDelete?.title}"?`}
+          onConfirm={() => handleDelete()}
+          onCancel={() => setProgramToDelete(null)}
+          destructive
+        />
+      </Modal>
     </div>
   );
 }
